tests: Move some base classes to conftest.py
This allows the sharing of some useful base classes to multiple test
files.
