+++ setup.cfg
@@ -12,13 +12,13 @@ python-tag = py3
 
 [pydocstyle]
 convention = numpy
-match = ^((?!\.sync-zenodo-metadata|setup).)*\.py$
+match = ^((?!\.sync-zenodo-metadata|setup|conftest).)*\.py$
 match-dir = ^((?!\.|tests|mistune).)*$
 ignore-decorators = "deprecated"
 
 [flake8]
 max-line-length = 100
-exclude = mistune
+exclude = mistune,conftest.py
 select = E,F,W
 ignore = E123,E126,E226,E241,E704,W503,W504
